Updates (#1)
* Added UserName parameter

Per user request this parameter was added so the Get-Credential username can be specified and not current user.

* FQDN and Shares added

Updated to ensure that the FQDN of the Exchange server is used for the connection. The script also creates a temporary file shares for uploading and downloading between the host and Exchange server.

* Improved error handling

Added error handling for the output path, script path, share creation, and remote PowerShell session.

* Timer added to end script

Added a timer so that the script doesn't wait indefinitely for results to be returned. Also modified the share to be hidden for added security.

* Zip results error handling added

Attempts to zip the entire directory. If that fails, create a zip file and attempt to add each file individually.

* Final zip update

Fix the path for the results zip

* Credential validation added

Added a credential test so that a mistyped password won't cause script to error out.

* Minor updates

Added parameter to specify where data resides and error handling for decompressing zip files.

Co-authored-by: jmartinmsft <72947063+jmartinmsft@users.noreply.github.com>